requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl_cffi==0.11.2
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==8.0.2
exceptiongroup==1.3.0
fastapi==0.115.12
flatbuffers==25.2.10
frozendict==2.4.6
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
groq==0.26.0
grpcio==1.72.1
h11==0.16.0
h5py==3.11.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.1
keras==3.10.0
libclang==18.1.1
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
multitasking==0.0.11
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
peewee==3.18.1
phidata==2.7.10
platformdirs==4.3.8
primp==0.15.0
protobuf==4.25.8
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.7.0
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.2.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yfinance==0.2.61
